skills/Organizer/Tools/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyPDF2
